Refactor default localisation instance creation from index.js to localisation.js
