+++ src/main/app/src/localisation.js
@@ -1,14 +1,16 @@
 import i18n from 'i18next';
 import XHR from 'i18next-xhr-backend';
-import { get, capitalize, toLower, isFunction } from 'lodash';
+import _ from 'lodash';
+import { getLocalisation } from './apiUtils';
+import getTranslations from './translations';
 
 const formatMap = {
-  toLower,
-  capitalize,
+  toLower: _.toLower,
+  capitalize: _.capitalize,
   unCapitalize: value => {
-    const low = toLower(value);
+    const low = _.toLower(value);
     // if second character changes, return original value
-    return get(value, 1) === get(low, 1) ? low : value;
+    return _.get(value, 1) === _.get(low, 1) ? low : value;
   },
 };
 
@@ -49,12 +51,46 @@ const createLocalisation = ({
     }),
     interpolation: {
       format(value, format, lng) {
-        return isFunction(formatMap[format]) ? formatMap[format](value) : value;
+        return _.isFunction(formatMap[format])
+          ? formatMap[format](value)
+          : value;
       },
     },
   });
 
   return instance;
 };
+
+const isDev = process.env.NODE_ENV === 'development';
+
+const createLocalisationLoader = ({ httpClient, apiUrls }) => async ({
+  namespace,
+  language,
+}) => {
+  const localisation = await getLocalisation({
+    category: namespace,
+    locale: language,
+    httpClient,
+    apiUrls,
+  });
+
+  const translations = getTranslations();
+
+  return _.get(translations, [language, namespace])
+    ? _.merge({}, translations[language][namespace], localisation || {})
+    : localisation;
+};
+
+let localisationInstance = null;
+
+export const createDefaultLocalisation = ({ httpClient, apiUrls }) => {
+  localisationInstance = createLocalisation({
+    debug: isDev,
+    loadLocalisation: createLocalisationLoader({ httpClient, apiUrls }),
+  });
+  return localisationInstance;
+};
+
+export const getDefaultLocalisation = () => localisationInstance;
 
 export default createLocalisation;
